+++ easy-retry-starters/easy-retry-starter-common/pom.xml
@@ -9,9 +9,9 @@
 		<relativePath>../pom.xml</relativePath>
 	</parent>
 
-	<artifactId>easy-retry-mybatis-starter</artifactId>
-	<name>easy-retry-mybatis-starter</name>
-	<description>easy-retry-mybatis-starter</description>
+	<artifactId>easy-retry-starter-common</artifactId>
+	<name>easy-retry-starter-common</name>
+	<description>easy-retry-starter-common</description>
 	<dependencies>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
@@ -26,28 +26,6 @@
 		<dependency>
 			<groupId>${project.groupId}</groupId>
 			<artifactId>easy-retry-core</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>${project.groupId}</groupId>
-			<artifactId>easy-retry-mybatis-extension</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>org.mybatis</groupId>
-			<artifactId>mybatis-spring</artifactId>
-			<version>2.0.6</version>
-		</dependency>
-
-		<dependency>
-			<groupId>org.springframework</groupId>
-			<artifactId>spring-tx</artifactId>
-			<version>5.1.12.RELEASE</version>
-		</dependency>
-
-		<dependency>
-			<groupId>${project.groupId}</groupId>
-			<artifactId>easy-retry-starter-common</artifactId>
 		</dependency>
 
 	</dependencies>
