+++ pom.xml
@@ -14,7 +14,7 @@
 		<java.version>1.8</java.version>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<revision>1.0.0.RC3</revision>
+		<revision>1.0.0.RC4-SNAPSHOT</revision>
 		<slf4j.version>1.7.7</slf4j.version>
 		<lombok.version>1.18.18</lombok.version>
 		<guava.version>16.0</guava.version>
@@ -140,6 +140,12 @@
 			<dependency>
 				<groupId>${project.groupId}</groupId>
 				<artifactId>easy-retry-mybatis-extension</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+
+			<dependency>
+				<groupId>${project.groupId}</groupId>
+				<artifactId>easy-retry-starter-common</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 		</dependencies>
